Syncing changes for release 0.5.0

+++ src/main/bash/commands/besman-list.sh
@@ -2,75 +2,78 @@
 
 function __bes_list {
 
-local flag=$1
-local env sorted_list
+    local flag=$1
+    local env sorted_list
 
-# For listing playbooks
-if [[ ( -n $flag ) && ( ( $flag == "--playbook" ) || ( $flag == "-P" ) ) ]]; then
+    # For listing playbooks
+    if [[ (-n $flag) && (($flag == "--playbook") || ($flag == "-P")) ]]; then
 
-    __besman_list_playbooks
+        __besman_list_playbooks
 
-elif [[ ( -n $flag ) && ( $flag == "--role" ) ]]; then
+    elif [[ (-n $flag) && ($flag == "--role") ]]; then
 
-    __besman_list_roles
-elif [[ ( -n $flag ) && ( ( $flag == "--environment" ) || ( $flag == "-env" ) ) ]]; then
+        __besman_list_roles
+    elif [[ (-n $flag) && (($flag == "--environment") || ($flag == "-env")) ]]; then
 
-    __besman_list_envs
+        __besman_list_envs
 
-else
-    
-    __besman_echo_white "---------------------------ENVIRONMENTS-----------------------------------------------"
-    __besman_echo_no_colour ""
-    __besman_list_envs
-    __besman_echo_no_colour ""
-    __besman_echo_white "---------------------------PLAYBOOKS--------------------------------------------------"
-    __besman_echo_no_colour ""
-    __besman_list_playbooks
-    __besman_echo_no_colour ""
-    __besman_echo_white "---------------------------ROLES------------------------------------------------------"
-    __besman_echo_no_colour ""
-    __besman_list_roles
-    __besman_echo_no_colour ""
-fi
+    else
+
+        __besman_echo_white "---------------------------ENVIRONMENTS-----------------------------------------------"
+        __besman_echo_no_colour ""
+        __besman_list_envs
+        __besman_echo_no_colour ""
+        __besman_echo_white "---------------------------PLAYBOOKS--------------------------------------------------"
+        __besman_echo_no_colour ""
+        __besman_list_playbooks
+        __besman_echo_no_colour ""
+        __besman_echo_white "---------------------------ROLES------------------------------------------------------"
+        __besman_echo_no_colour ""
+        __besman_list_roles
+        __besman_echo_no_colour ""
+    fi
 }
-function __besman_list_envs()
-{
-    local current_version current_env installed_annotation remote_annotation
+function __besman_list_envs() {
+    local current_version current_env installed_annotation remote_annotation local_list
+
+    local_list="$BESMAN_DIR/var/list.txt"
+
     __besman_check_repo_exist || return 1
-    __besman_update_list
+    __besman_update_list || return 1
     # __besman_echo_no_colour "Github Org    Repo                             Environment     Version"
     # __besman_echo_no_colour "-----------------------------------------------------------------------------------"
 
-    [[ -f "$BESMAN_DIR/var/current" ]] &&  current_env=$(cat "$BESMAN_DIR/var/current")
+    [[ -f "$BESMAN_DIR/var/current" ]] && current_env=$(cat "$BESMAN_DIR/var/current")
     [[ -f "$BESMAN_DIR/envs/besman-$current_env/current" ]] && current_version=$(cat "$BESMAN_DIR/envs/besman-$current_env/current")
 
     installed_annotation=$(__besman_echo_red "*")
     remote_annotation=$(__besman_echo_yellow "^")
-    
-    # For listing environments
-    printf "%-14s %-32s %-25s %-8s\n" "Github Org" "Repo" "Environment" "Version"
-    __besman_echo_no_colour "-----------------------------------------------------------------------------------"
 
-    
-    sed -i '/^$/d' "$BESMAN_DIR/var/list.txt"
-    sorted_list=$(sort "$BESMAN_DIR/var/list.txt")
-    echo "$sorted_list" > "$BESMAN_DIR/var/list.txt"
+    # For listing environments
+    printf "%-25s %15s %15s\n" "Environment" "Author" "Version"
+    __besman_echo_no_colour "-----------------------------------------------------------"
+
+    [[ -f "$BESMAN_DIR/tmp/environment_details.txt" ]] && cp "$BESMAN_DIR/tmp/environment_details.txt" "$local_list" && rm "$BESMAN_DIR/tmp/environment_details.txt"
+
+    sed -i '/^$/d' "$local_list"
+    sorted_list=$(sort "$local_list")
+    echo "$sorted_list" >"$local_list"
     OLD_IFS=$IFS
-    IFS="/"
-      
-    while read -r line; 
-    do 
-        converted_line=$(echo "$line" | sed 's|,|/|g')
-        read -r org repo env version <<< "$converted_line"
-        if [[ ("$env" == "$current_env") && ("$version" == "$current_version") ]] 
-        then
-            printf "%-14s %-32s %-25s %-8s\n" "$org" "$repo" "$env" "$version$installed_annotation"
+    IFS=" "
+
+    while read -r line; do
+        # converted_line=$(echo "$line" | sed 's|,|/|g')
+        read -r env author version <<<"$line"
+
+        # echo "line=$line"
+        if [[ ("$env" == "$current_env") && ("$version" == "$current_version") ]]; then
+            printf "%-25s %15s %25s\n" "$env" "$author" "$version$installed_annotation"
         else
-            printf "%-14s %-32s %-25s %-8s\n" "$org" "$repo" "$env" "$version$remote_annotation"
-            
+            printf "%-25s %15s %25s\n" "$env" "$author" "$version$remote_annotation"
+
         fi
-        
-    done < "$BESMAN_DIR/var/list.txt"
+
+    done <"$local_list"
     IFS=$OLD_IFS
 
     __besman_echo_no_colour ""
@@ -85,26 +88,29 @@ function __besman_list_envs()
 
     if [[ $BESMAN_LOCAL_ENV == "true" ]]; then
 
-        __besman_echo_yellow "Pointing to local dir $BESMAN_LOCAL_ENV_DIR"
+        __besman_echo_white "Listing from local dir $(__besman_echo_yellow "$BESMAN_LOCAL_ENV_DIR")"
         __besman_echo_no_colour ""
-        __besman_echo_white "If you wish to change, run the below command"
+        __besman_echo_white "If you wish to list from remote repo, run the below command"
         __besman_echo_yellow "$ bes set BESMAN_LOCAL_ENV false"
-        __besman_echo_yellow "$ bes set BESMAN_ENV_REPOS <GitHub Org>"
-    else      
-        __besman_echo_yellow "Pointing to $BESMAN_ENV_REPOS; tree - $BESMAN_ENV_REPO_BRANCH"
-        __besman_echo_yellow "If you wish to change the repo run the below command"
-        __besman_echo_yellow "$ bes set BESMAN_ENV_REPOS <GitHub Org>"
+        __besman_echo_yellow "$ bes set BESMAN_ENV_REPO <GitHub Org>/<Repo name>"
+    else
+        __besman_echo_white "Listing from $(__besman_echo_yellow "$BESMAN_ENV_REPO"); branch - $(__besman_echo_yellow "$BESMAN_ENV_REPO_BRANCH")"
+        __besman_echo_no_colour ""
+        __besman_echo_white "If you wish to change the repo, run the below command"
+        __besman_echo_yellow "$ bes set BESMAN_ENV_REPO <GitHub Org>/<Repo name>"
+        __besman_echo_no_colour ""
+        __besman_echo_white "If you wish to change the branch, run the below command"
+        __besman_echo_yellow "$ bes set BESMAN_ENV_REPO_BRANCH <branch>/<tag>"
     fi
 }
-function __besman_check_repo_exist()
-{
+function __besman_check_repo_exist() {
     local namespace repo response repo_url
     [[ $BESMAN_LOCAL_ENV == "true" ]] && return 0
-    namespace=$(echo "$BESMAN_ENV_REPOS" | cut -d "/" -f 1)
-    repo=$(echo "$BESMAN_ENV_REPOS" | cut -d "/" -f 2)
+    namespace=$(echo "$BESMAN_ENV_REPO" | cut -d "/" -f 1)
+    repo=$(echo "$BESMAN_ENV_REPO" | cut -d "/" -f 2)
     repo_url="https://api.github.com/repos/$namespace/$repo"
 
-    response=$(curl --head --silent "$repo_url" | head -n 1 | awk '{print $2}')
+    response=$(curl --head --insecure --silent "$repo_url" | head -n 1 | awk '{print $2}')
 
     if [ "$response" -ne 200 ]; then
         __besman_echo_red "Repository $repo does not exist under $namespace"
@@ -113,42 +119,72 @@ function __besman_check_repo_exist()
 
 }
 
-function __besman_update_list()
-{
-    local bes_list
-    if [[ ( -n $BESMAN_LOCAL_ENV ) && ( $BESMAN_LOCAL_ENV == "true" )]]; then
+function __besman_update_list() {
+    local bes_list exit_code
+    if [[ (-n $BESMAN_LOCAL_ENV) && ($BESMAN_LOCAL_ENV == "true") ]]; then
         local env_dir_list bes_list
-        [[ -z $BESMAN_LOCAL_ENV_DIR ]] && __besman_echo_red "Please set the local env dir first" && return 1
-        [[ ! -d $BESMAN_LOCAL_ENV_DIR ]] && __besman_echo_red "Could not find dir $BESMAN_LOCAL_ENV_DIR" && return 1
+        if [[ -z $BESMAN_LOCAL_ENV_DIR ]]; then
+            __besman_echo_red "Could not find your local environment dir"
+            __besman_echo_no_colour ""
+            __besman_echo_white "Use the below command to set it first"
+            __besman_echo_no_colour ""
+            __besman_echo_yellow "$ bes set BESMAN_LOCAL_ENV_DIR <complete path to your local env dir>"
+            __besman_echo_no_colour ""
 
-        env_dir_list=$(< "$BESMAN_LOCAL_ENV_DIR/list.txt")
-        bes_list=$BESMAN_DIR/var/list.txt
-        echo "$env_dir_list" > "$bes_list"
+            return 1
+        fi
+        [[ ! -d $BESMAN_LOCAL_ENV_DIR ]] && __besman_echo_red "Could not find dir $BESMAN_LOCAL_ENV_DIR" && return 1
+        local env_script_file="$BESMAN_DIR/scripts/besman-get-env-list.py"
+        [[ ! -f $env_script_file ]] && __besman_echo_red "Could not find script file $env_script_file" && return 1
+
+        python3 $env_script_file
+        # env_dir_list=$(< "$BESMAN_LOCAL_ENV_DIR/list.txt")
+        # bes_list=$BESMAN_DIR/var/list.txt
+        # echo "$env_dir_list" > "$bes_list"
     else
-            
-        local org repo path 
-        org=$(echo "$BESMAN_ENV_REPOS" | cut -d "/" -f 1)
-        repo=$(echo "$BESMAN_ENV_REPOS" | cut -d "/" -f 2)
-        branch=$BESMAN_ENV_REPO_BRANCH
-        bes_list="$BESMAN_DIR/var/list.txt"
-        path="https://raw.githubusercontent.com/$org/$repo/$branch/list.txt"
-        __besman_secure_curl "$path" > "$bes_list"
+
+        # local org repo path
+        # org=$(echo "$BESMAN_ENV_REPO" | cut -d "/" -f 1)
+        # repo=$(echo "$BESMAN_ENV_REPO" | cut -d "/" -f 2)
+        # branch=$BESMAN_ENV_REPO_BRANCH
+        # bes_list="$BESMAN_DIR/var/list.txt"
+        # # path="https://raw.githubusercontent.com/$org/$repo/$branch/list.txt"
+        # # __besman_secure_curl "$path" > "$bes_list"
+        local env_script_file="$BESMAN_DIR/scripts/besman-get-env-list.py"
+        [[ ! -f $env_script_file ]] && __besman_echo_red "Could not find script file $env_script_file" && return 1
+
+        python3 $env_script_file
+
+        exit_code=$?
+
+        if [[ $exit_code -eq 0 ]]; then
+            return 0
+        elif [[ $exit_code -eq 1 ]]; then
+            __besman_echo_red "Error fetching data."
+            return 1
+        elif [[ $exit_code -eq 2 ]]; then
+            __besman_echo_red "Error parsing JSON."
+            return 1
+        elif [[ $exit_code -eq 3 ]]; then
+            __besman_echo_red "Error writing to file."
+            return 1
+        else
+            __besman_echo_red "An unexpected error occurred."
+            return 1
+        fi
     fi
 
 }
 
 # Function to extract repository names from a JSON response
-function __besman_extract_repo_names()
-{
-  echo "$1" | grep -oP '"full_name": "\K[^"]+'
+function __besman_extract_repo_names() {
+    echo "$1" | grep -oP '"full_name": "\K[^"]+'
 }
 
-
-function __besman_list_roles()
-{
+function __besman_list_roles() {
     local api_url repo_names all_repo_names page_num ansible_roles
 
-        if [[ -z "$BESMAN_GH_TOKEN" ]]; then
+    if [[ -z "$BESMAN_GH_TOKEN" ]]; then
         __besman_echo_yellow "Github token missing. Please use the below command to export the token"
         __besman_echo_no_colour ""
         __besman_echo_no_colour "$ bes set BESMAN_GH_TOKEN <copied token>"
@@ -174,81 +210,102 @@ function __besman_list_roles()
     done
 
     ansible_roles=$(echo "$all_repo_names" | grep "ansible-role-*")
-    
+
     printf "%-14s %10s \n" "Github Org" "Repo"
     __besman_echo_no_colour "-----------------------------------"
-    for i in $ansible_roles
-    do
+    for i in $ansible_roles; do
         converted_i=$(echo "$i" | sed "s|/| |g")
-        read -r org repo <<< "$converted_i"
+        read -r org repo <<<"$converted_i"
         printf "%-14s %-32s \n" "$org" "$repo"
     done
-    
 
 }
 
-function __besman_get_playbook_details()
-{
+function __besman_get_playbook_details() {
     local scripts_file
-
+    local environment=$1
+    local version=$2
     scripts_file="$BESMAN_DIR/scripts/besman-get-playbook-details.py"
 
     [[ ! -f "$scripts_file" ]] && __besman_echo_red "Could not find $scripts_file" && return 1
 
-    python3 "$scripts_file"
+    if [[ -z $environment || -z $version ]]; then
+        python3 "$scripts_file" --master_list True
+    else
+        python3 "$scripts_file" --environment "$environment" --version "$version"
+    fi
+
+    if [[ "$?" != "0" ]]; then
+        __besman_echo_red "Error while fetching playbook details"
+        return 1
+    fi
 }
-
-function __besman_list_playbooks()
-{
-
+function __besman_list_playbooks() {
 
     local playbook_details_file playbook_details local_annotation remote_annotation
 
+    [[ ! -f "$BESMAN_DIR/var/current" || -z $(cat "$BESMAN_DIR/var/current") ]] && __besman_echo_red "Missing environment" && __besman_echo_white "\nInstall an environment to get the list of compatible playbooks" && return 1
+
+    local current_env=$(cat "$BESMAN_DIR/var/current")
+
+    [[ -z $current_env ]] && __besman_echo_red "Could not find installed environment" && return 1
+
+    [[ ! -d "$BESMAN_DIR/envs/besman-$current_env" ]] && __besman_echo_red "Could not find installed environment" && return 1
+
+    local current_env_version=$(cat "$BESMAN_DIR/envs/besman-$current_env/current")
+
     playbook_details_file="$BESMAN_DIR/tmp/playbook_details.txt"
 
-    __besman_get_playbook_details || return 1
+    __besman_get_playbook_details "$current_env" "$current_env_version" || return 1
 
     playbook_details=$(cat "$playbook_details_file")
 
-    [[ ( ! -f "$playbook_details_file" ) || ( -z $playbook_details ) ]] && __besman_echo_red "Could not find playbook details" && return 1
+    [[ (! -f "$playbook_details_file") || (-z $playbook_details) ]] && __besman_echo_red "Could not find playbook details" && return 1
 
     local_annotation=$(__besman_echo_red "+")
     remote_annotation=$(__besman_echo_yellow "^")
-    
+    printf "\n%-15s Compatible playbooks for $current_env $current_env_version"
+    __besman_echo_white "\n=======================================================================\n"
     printf "%-25s %-10s %-15s %-8s\n" "Name" "Version" "Type" "Author"
     __besman_echo_no_colour "----------------------------------------------------------------------"
 
     OLD_IFS=$IFS
     IFS=" "
-      
-    while read -r line; 
-    do 
+
+    while read -r line; do
         # converted_line=$(echo "$line" | sed 's|,|/|g')
-        read -r name version type author <<< "$line"
-        if [[ -f "$BESMAN_PLAYBOOK_DIR/besman-$name-$version-playbook.sh" ]] 
-        then
-            
+        read -r name version type author <<<"$line"
+        if [[ -f "$BESMAN_PLAYBOOK_DIR/besman-$name-$version-playbook.sh" ]]; then
+
             printf "%-25s %-10s %-15s %-8s\n" "$name" "$version" "$type" "$author$local_annotation"
         else
             printf "%-25s %-10s %-15s %-8s\n" "$name" "$version" "$type" "$author$remote_annotation"
 
         fi
-        
-    done <<< "$playbook_details"
+
+    done <<<"$playbook_details"
     IFS=$OLD_IFS
 
-    __besman_echo_no_colour ""
-    __besman_echo_no_colour "======================================================================="
-    __besman_echo_no_colour "$remote_annotation - remote playbook"
-    __besman_echo_no_colour "$local_annotation - local playbook"
-    __besman_echo_no_colour "======================================================================="
-    __besman_echo_no_colour ""
+    if [[ $BESMAN_LOCAL_ENV == "false" ]]; then
 
-    __besman_echo_no_colour ""
-    __besman_echo_yellow "Pointing to $BESMAN_PLAYBOOK_REPO; tree - $BESMAN_PLAYBOOK_REPO_BRANCH"
-    __besman_echo_yellow "If you wish to change the repo run the below command"
-    __besman_echo_yellow "$ bes set BESMAN_PLAYBOOK_REPO <GitHub Org>/<Repo name>"
+        __besman_echo_no_colour ""
+        __besman_echo_no_colour "======================================================================="
+        __besman_echo_no_colour "$remote_annotation - remote playbook"
+        __besman_echo_no_colour "$local_annotation - local playbook"
+        __besman_echo_no_colour "======================================================================="
+        __besman_echo_no_colour ""
 
-    [[ -f $playbook_details_file ]] && rm "$playbook_details_file"
+        __besman_echo_no_colour ""
+        __besman_echo_white "Listing from $(__besman_echo_yellow "$BESMAN_PLAYBOOK_REPO"); branch - $(__besman_echo_yellow "$BESMAN_PLAYBOOK_REPO_BRANCH")"
+        __besman_echo_white "If you wish to change the repo run the below command"
+        __besman_echo_yellow "$ bes set BESMAN_PLAYBOOK_REPO <GitHub Org>/<Repo name>"
+        __besman_echo_no_colour ""
+        __besman_echo_white "If you wish to change the branch run the below command"
+        __besman_echo_yellow "$ bes set BESMAN_PLAYBOOK_REPO_BRANCH <branch>/<tag>"
+
+        [[ -f $playbook_details_file ]] && rm "$playbook_details_file"
+    else
+        __besman_echo_white "\n"
+    fi
 
 }
